Making Dia operation common for netcore, & uwp
