+++ src/Adapter/PlatformServices.Shared/netstandard1.0/Services/ns10FileOperations.cs
@@ -16,12 +16,6 @@ namespace Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices
     /// </summary>
     public class FileOperations : IFileOperations
     {
-        private static MethodInfo methodGetNavigationData;
-        private static PropertyInfo propertyFileName;
-        private static PropertyInfo propertyMinLineNumber;
-        private static Type typeDiaSession;
-        private static Type typeDiaNavigationData;
-
         /// <summary>
         /// Initializes static members of the <see cref="FileOperations"/> class.
         /// </summary>
@@ -82,14 +76,7 @@ namespace Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices
         /// <returns> A Navigation session instance for the current platform. </returns>
         public object CreateNavigationSession(string source)
         {
-            // Create instance only when DiaSession is found in Object Model.
-            if (typeDiaSession != null && typeDiaNavigationData != null)
-            {
-                var messageFormatOnException = string.Join("MSTestDiscoverer:DiaSession: Could not create diaSession for source:", source, ". Reason:{0}");
-                return SafeInvoke(() => Activator.CreateInstance(typeDiaSession, source), messageFormatOnException);
-            }
-
-            return null;
+            return DiaSessionOperations.CreateNavigationSession(source);
         }
 
         /// <summary>
@@ -102,22 +89,7 @@ namespace Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices
         /// <param name="fileName"> The file name. </param>
         public void GetNavigationData(object navigationSession, string className, string methodName, out int minLineNumber, out string fileName)
         {
-            // Set default values.
-            fileName = null;
-            minLineNumber = -1;
-
-            // Get navigation data only when DiaSession is found in Object Model.
-            if (typeDiaSession != null && typeDiaNavigationData != null)
-            {
-                var messageFormatOnException = string.Join("MSTestDiscoverer:DiaSession: Could not get navigation data for class:", className, ". Reason:{0}");
-                var data = SafeInvoke(() => methodGetNavigationData.Invoke(navigationSession, new object[] { className, methodName }), messageFormatOnException);
-
-                if (data != null)
-                {
-                    fileName = (string)propertyFileName?.GetValue(data);
-                    minLineNumber = (int)(propertyMinLineNumber?.GetValue(data) ?? -1);
-                }
-            }
+            DiaSessionOperations.GetNavigationData(navigationSession, className, methodName, out minLineNumber, out fileName);
         }
 
         /// <summary>
@@ -126,8 +98,7 @@ namespace Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices
         /// <param name="navigationSession"> The navigation session. </param>
         public void DisposeNavigationSession(object navigationSession)
         {
-            var diaSession = navigationSession as IDisposable;
-            diaSession?.Dispose();
+            DiaSessionOperations.DisposeNavigationSession(navigationSession);
         }
 
         /// <summary>
@@ -148,29 +119,7 @@ namespace Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices
         /// <param name="diaNavigationData">Type name of DiaNavigationData class.</param>
         internal static void Initialize(string diaSession, string diaNavigationData)
         {
-            typeDiaSession = Type.GetType(diaSession, false);
-            typeDiaNavigationData = Type.GetType(diaNavigationData, false);
-
-            if (typeDiaSession != null && typeDiaNavigationData != null)
-            {
-                methodGetNavigationData = typeDiaSession.GetRuntimeMethod("GetNavigationData", new[] { typeof(string), typeof(string) });
-                propertyFileName = typeDiaNavigationData.GetRuntimeProperty("FileName");
-                propertyMinLineNumber = typeDiaNavigationData.GetRuntimeProperty("MinLineNumber");
-            }
-        }
-
-        private static object SafeInvoke<T>(Func<T> action, string messageFormatOnException = null)
-        {
-            try
-            {
-                return action.Invoke();
-            }
-            catch (Exception)
-            {
-                // todo : Add EqtTrace
-            }
-
-            return null;
+            DiaSessionOperations.Initialize(diaSession, diaNavigationData);
         }
     }
