+++ src/Adapter/PlatformServices.Universal/PlatformServices.Universal.csproj
@@ -119,6 +119,9 @@
     <Compile Include="..\PlatformServices.Shared\netstandard1.0\Services\ns10ReflectionOperations.cs">
       <Link>Services\ns10ReflectionOperations.cs</Link>
     </Compile>
+    <Compile Include="..\PlatformServices.Shared\netstandard1.0\Services\ns10DiaSessionOperations.cs">
+      <Link>Services\ns10DiaSessionOperations.cs</Link>
+    </Compile>
     <Compile Include="..\PlatformServices.Shared\netstandard1.0\Services\ns10SettingsProvider.cs">
       <Link>Services\ns10SettingsProvider.cs</Link>
     </Compile>
@@ -143,6 +146,7 @@
     <Compile Include="..\PlatformServices.Shared\netstandard1.0\Services\ns10TraceListenerManager.cs">
       <Link>Services\ns10TraceListenerManager.cs</Link>
     </Compile>
+    <Compile Include="Friends.cs" />
     <Compile Include="Services\UniversalAdapterTraceLogger.cs" />
     <Compile Include="Services\UniversalFileOperations.cs" />
     <Compile Include="Services\UniversalTestSource.cs" />
