+++ test/UnitTests/PlatformServices.Universal.Unit.Tests/PlatformServices.Universal.Unit.Tests.csproj
@@ -60,6 +60,9 @@
     <Compile Include="..\PlatformServices.Shared.Unit.Tests\netstandard1.0\ns10AssemblyAttributes.cs">
       <Link>Services\ns10AssemblyAttributes.cs</Link>
     </Compile>
+    <Compile Include="..\PlatformServices.Shared.Unit.Tests\netstandard1.0\ns10DiaSessionOperationsTests.cs">
+      <Link>Services\ns10DiaSessionOperationsTests.cs</Link>
+    </Compile>
     <Compile Include="..\PlatformServices.Shared.Unit.Tests\netstandard1.0\ns10ReflectionOperationsTests.cs">
       <Link>Services\ns10ReflectionOperationsTests.cs</Link>
     </Compile>
